# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cafe QR Coupon System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CreateBatchBodyDiscountType =
  (typeof CreateBatchBodyDiscountType)[keyof typeof CreateBatchBodyDiscountType];

export const CreateBatchBodyDiscountType = {
  percentage: "percentage",
  fixed: "fixed",
  free_item: "free_item",
} as const;

export type CreateBatchBodyLanguage =
  (typeof CreateBatchBodyLanguage)[keyof typeof CreateBatchBodyLanguage];

export const CreateBatchBodyLanguage = {
  en: "en",
  tr: "tr",
} as const;

export interface CreateBatchBody {
  /**
   * @minimum 1
   * @maximum 100
   */
  quantity: number;
  discountType: CreateBatchBodyDiscountType;
  discountValue?: number;
  freeItemName?: string;
  displayTitle: string;
  displaySubtitle?: string;
  language?: CreateBatchBodyLanguage;
  backgroundColor?: string;
  logoUrl?: string;
  footerText?: string;
  expiresAt?: string;
}

export type BatchDiscountType =
  (typeof BatchDiscountType)[keyof typeof BatchDiscountType];

export const BatchDiscountType = {
  percentage: "percentage",
  fixed: "fixed",
  free_item: "free_item",
} as const;

export interface Batch {
  id: number;
  quantity: number;
  discountType: BatchDiscountType;
  discountValue?: number;
  freeItemName?: string;
  displayTitle: string;
  displaySubtitle?: string;
  language?: string;
  backgroundColor?: string;
  logoUrl?: string;
  footerText?: string;
  expiresAt?: string;
  createdAt: string;
  activeCount: number;
  usedCount: number;
  expiredCount: number;
}

export type BatchWithCouponsDiscountType =
  (typeof BatchWithCouponsDiscountType)[keyof typeof BatchWithCouponsDiscountType];

export const BatchWithCouponsDiscountType = {
  percentage: "percentage",
  fixed: "fixed",
  free_item: "free_item",
} as const;

export type CouponDiscountType =
  (typeof CouponDiscountType)[keyof typeof CouponDiscountType];

export const CouponDiscountType = {
  percentage: "percentage",
  fixed: "fixed",
  free_item: "free_item",
} as const;

export type CouponStatus = (typeof CouponStatus)[keyof typeof CouponStatus];

export const CouponStatus = {
  active: "active",
  used: "used",
  expired: "expired",
} as const;

export interface Coupon {
  id: number;
  batchId: number;
  code: string;
  discountType: CouponDiscountType;
  discountValue?: number;
  freeItemName?: string;
  displayTitle: string;
  displaySubtitle?: string;
  language?: string;
  backgroundColor?: string;
  logoUrl?: string;
  footerText?: string;
  isUsed: boolean;
  createdAt: string;
  redeemedAt?: string;
  expiresAt?: string;
  status: CouponStatus;
}

export interface BatchWithCoupons {
  id: number;
  quantity: number;
  discountType: BatchWithCouponsDiscountType;
  discountValue?: number;
  freeItemName?: string;
  displayTitle: string;
  displaySubtitle?: string;
  language?: string;
  backgroundColor?: string;
  logoUrl?: string;
  footerText?: string;
  expiresAt?: string;
  createdAt: string;
  coupons: Coupon[];
}

export type CouponPublicDiscountType =
  (typeof CouponPublicDiscountType)[keyof typeof CouponPublicDiscountType];

export const CouponPublicDiscountType = {
  percentage: "percentage",
  fixed: "fixed",
  free_item: "free_item",
} as const;

export type CouponPublicStatus =
  (typeof CouponPublicStatus)[keyof typeof CouponPublicStatus];

export const CouponPublicStatus = {
  active: "active",
  used: "used",
  expired: "expired",
} as const;

export interface CouponPublic {
  code: string;
  discountType: CouponPublicDiscountType;
  discountValue?: number;
  freeItemName?: string;
  displayTitle: string;
  displaySubtitle?: string;
  language?: string;
  backgroundColor?: string;
  logoUrl?: string;
  footerText?: string;
  status: CouponPublicStatus;
  expiresAt?: string;
}

export interface CouponValidation {
  valid: boolean;
  coupon?: Coupon;
  message: string;
}

export interface DashboardSummary {
  totalBatches: number;
  totalCoupons: number;
  activeCoupons: number;
  usedCoupons: number;
  expiredCoupons: number;
  redemptionRate: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  created: "created",
  redeemed: "redeemed",
  expired: "expired",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  couponCode: string;
  batchTitle: string;
  timestamp: string;
}

export type ListCouponsParams = {
  batchId?: number;
  status?: ListCouponsStatus;
};

export type ListCouponsStatus =
  (typeof ListCouponsStatus)[keyof typeof ListCouponsStatus];

export const ListCouponsStatus = {
  active: "active",
  used: "used",
  expired: "expired",
} as const;