# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cafe QR Coupon System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all coupon batches
 */
export const ListBatchesResponseItem = zod.object({
  id: zod.number(),
  quantity: zod.number(),
  discountType: zod.enum(["percentage", "fixed", "free_item"]),
  discountValue: zod.number().optional(),
  freeItemName: zod.string().optional(),
  displayTitle: zod.string(),
  displaySubtitle: zod.string().optional(),
  language: zod.string().optional(),
  backgroundColor: zod.string().optional(),
  logoUrl: zod.string().optional(),
  footerText: zod.string().optional(),
  expiresAt: zod.coerce.date().optional(),
  createdAt: zod.coerce.date(),
  activeCount: zod.number(),
  usedCount: zod.number(),
  expiredCount: zod.number(),
});
export const ListBatchesResponse = zod.array(ListBatchesResponseItem);

/**
 * @summary Create a new coupon batch with coupons
 */
export const createBatchBodyQuantityMax = 100;

export const CreateBatchBody = zod.object({
  quantity: zod.number().min(1).max(createBatchBodyQuantityMax),
  discountType: zod.enum(["percentage", "fixed", "free_item"]),
  discountValue: zod.number().optional(),
  freeItemName: zod.string().optional(),
  displayTitle: zod.string(),
  displaySubtitle: zod.string().optional(),
  language: zod.enum(["en", "tr"]).optional(),
  backgroundColor: zod.string().optional(),
  logoUrl: zod.string().optional(),
  footerText: zod.string().optional(),
  expiresAt: zod.coerce.date().optional(),
});

/**
 * @summary Get a single batch with its coupons
 */
export const GetBatchParams = zod.object({
  batchId: zod.coerce.number(),
});

export const GetBatchResponse = zod.object({
  id: zod.number(),
  quantity: zod.number(),
  discountType: zod.enum(["percentage", "fixed", "free_item"]),
  discountValue: zod.number().optional(),
  freeItemName: zod.string().optional(),
  displayTitle: zod.string(),
  displaySubtitle: zod.string().optional(),
  language: zod.string().optional(),
  backgroundColor: zod.string().optional(),
  logoUrl: zod.string().optional(),
  footerText: zod.string().optional(),
  expiresAt: zod.coerce.date().optional(),
  createdAt: zod.coerce.date(),
  coupons: zod.array(
    zod.object({
      id: zod.number(),
      batchId: zod.number(),
      code: zod.string(),
      discountType: zod.enum(["percentage", "fixed", "free_item"]),
      discountValue: zod.number().optional(),
      freeItemName: zod.string().optional(),
      displayTitle: zod.string(),
      displaySubtitle: zod.string().optional(),
      language: zod.string().optional(),
      backgroundColor: zod.string().optional(),
      logoUrl: zod.string().optional(),
      footerText: zod.string().optional(),
      isUsed: zod.boolean(),
      createdAt: zod.coerce.date(),
      redeemedAt: zod.coerce.date().optional(),
      expiresAt: zod.coerce.date().optional(),
      status: zod.enum(["active", "used", "expired"]),
    }),
  ),
});

/**
 * @summary List all coupons with optional filters
 */
export const ListCouponsQueryParams = zod.object({
  batchId: zod.coerce.number().optional(),
  status: zod.enum(["active", "used", "expired"]).optional(),
});

export const ListCouponsResponseItem = zod.object({
  id: zod.number(),
  batchId: zod.number(),
  code: zod.string(),
  discountType: zod.enum(["percentage", "fixed", "free_item"]),
  discountValue: zod.number().optional(),
  freeItemName: zod.string().optional(),
  displayTitle: zod.string(),
  displaySubtitle: zod.string().optional(),
  language: zod.string().optional(),
  backgroundColor: zod.string().optional(),
  logoUrl: zod.string().optional(),
  footerText: zod.string().optional(),
  isUsed: zod.boolean(),
  createdAt: zod.coerce.date(),
  redeemedAt: zod.coerce.date().optional(),
  expiresAt: zod.coerce.date().optional(),
  status: zod.enum(["active", "used", "expired"]),
});
export const ListCouponsResponse = zod.array(ListCouponsResponseItem);

/**
 * @summary Get coupon details by unique code (public)
 */
export const GetCouponByCodeParams = zod.object({
  code: zod.coerce.string(),
});

export const GetCouponByCodeResponse = zod.object({
  code: zod.string(),
  discountType: zod.enum(["percentage", "fixed", "free_item"]),
  discountValue: zod.number().optional(),
  freeItemName: zod.string().optional(),
  displayTitle: zod.string(),
  displaySubtitle: zod.string().optional(),
  language: zod.string().optional(),
  backgroundColor: zod.string().optional(),
  logoUrl: zod.string().optional(),
  footerText: zod.string().optional(),
  status: zod.enum(["active", "used", "expired"]),
  expiresAt: zod.coerce.date().optional(),
});

/**
 * @summary Validate a coupon for staff (shows full details)
 */
export const ValidateCouponParams = zod.object({
  code: zod.coerce.string(),
});

export const ValidateCouponResponse = zod.object({
  valid: zod.boolean(),
  coupon: zod
    .object({
      id: zod.number(),
      batchId: zod.number(),
      code: zod.string(),
      discountType: zod.enum(["percentage", "fixed", "free_item"]),
      discountValue: zod.number().optional(),
      freeItemName: zod.string().optional(),
      displayTitle: zod.string(),
      displaySubtitle: zod.string().optional(),
      language: zod.string().optional(),
      backgroundColor: zod.string().optional(),
      logoUrl: zod.string().optional(),
      footerText: zod.string().optional(),
      isUsed: zod.boolean(),
      createdAt: zod.coerce.date(),
      redeemedAt: zod.coerce.date().optional(),
      expiresAt: zod.coerce.date().optional(),
      status: zod.enum(["active", "used", "expired"]),
    })
    .optional(),
  message: zod.string(),
});

/**
 * @summary Redeem a coupon (staff only)
 */
export const RedeemCouponParams = zod.object({
  code: zod.coerce.string(),
});

export const RedeemCouponResponse = zod.object({
  id: zod.number(),
  batchId: zod.number(),
  code: zod.string(),
  discountType: zod.enum(["percentage", "fixed", "free_item"]),
  discountValue: zod.number().optional(),
  freeItemName: zod.string().optional(),
  displayTitle: zod.string(),
  displaySubtitle: zod.string().optional(),
  language: zod.string().optional(),
  backgroundColor: zod.string().optional(),
  logoUrl: zod.string().optional(),
  footerText: zod.string().optional(),
  isUsed: zod.boolean(),
  createdAt: zod.coerce.date(),
  redeemedAt: zod.coerce.date().optional(),
  expiresAt: zod.coerce.date().optional(),
  status: zod.enum(["active", "used", "expired"]),
});

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardSummaryResponse = zod.object({
  totalBatches: zod.number(),
  totalCoupons: zod.number(),
  activeCoupons: zod.number(),
  usedCoupons: zod.number(),
  expiredCoupons: zod.number(),
  redemptionRate: zod.number(),
});

/**
 * @summary Get recent coupon activity
 */
export const GetRecentActivityResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum(["created", "redeemed", "expired"]),
  couponCode: zod.string(),
  batchTitle: zod.string(),
  timestamp: zod.coerce.date(),
});
export const GetRecentActivityResponse = zod.array(
  GetRecentActivityResponseItem,
);