# lib/api-zod/src/generated/types/batchWithCoupons.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cafe QR Coupon System API
 * OpenAPI spec version: 0.1.0
 */
import type { BatchWithCouponsDiscountType } from "./batchWithCouponsDiscountType";
import type { Coupon } from "./coupon";

export interface BatchWithCoupons {
  id: number;
  quantity: number;
  discountType: BatchWithCouponsDiscountType;
  discountValue?: number;
  freeItemName?: string;
  displayTitle: string;
  displaySubtitle?: string;
  language?: string;
  backgroundColor?: string;
  logoUrl?: string;
  footerText?: string;
  expiresAt?: Date;
  createdAt: Date;
  coupons: Coupon[];
}