# lib/api-zod/src/generated/types/coupon.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cafe QR Coupon System API
 * OpenAPI spec version: 0.1.0
 */
import type { CouponDiscountType } from "./couponDiscountType";
import type { CouponStatus } from "./couponStatus";

export interface Coupon {
  id: number;
  batchId: number;
  code: string;
  discountType: CouponDiscountType;
  discountValue?: number;
  freeItemName?: string;
  displayTitle: string;
  displaySubtitle?: string;
  language?: string;
  backgroundColor?: string;
  logoUrl?: string;
  footerText?: string;
  isUsed: boolean;
  createdAt: Date;
  redeemedAt?: Date;
  expiresAt?: Date;
  status: CouponStatus;
}