# lib/api-zod/src/generated/types/couponPublic.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cafe QR Coupon System API
 * OpenAPI spec version: 0.1.0
 */
import type { CouponPublicDiscountType } from "./couponPublicDiscountType";
import type { CouponPublicStatus } from "./couponPublicStatus";

export interface CouponPublic {
  code: string;
  discountType: CouponPublicDiscountType;
  discountValue?: number;
  freeItemName?: string;
  displayTitle: string;
  displaySubtitle?: string;
  language?: string;
  backgroundColor?: string;
  logoUrl?: string;
  footerText?: string;
  status: CouponPublicStatus;
  expiresAt?: Date;
}