# lib/api-zod/src/generated/types/couponValidation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cafe QR Coupon System API
 * OpenAPI spec version: 0.1.0
 */
import type { Coupon } from "./coupon";

export interface CouponValidation {
  valid: boolean;
  coupon?: Coupon;
  message: string;
}