# lib/api-zod/src/generated/types/createBatchBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cafe QR Coupon System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBatchBodyDiscountType } from "./createBatchBodyDiscountType";
import type { CreateBatchBodyLanguage } from "./createBatchBodyLanguage";

export interface CreateBatchBody {
  /**
   * @minimum 1
   * @maximum 100
   */
  quantity: number;
  discountType: CreateBatchBodyDiscountType;
  discountValue?: number;
  freeItemName?: string;
  displayTitle: string;
  displaySubtitle?: string;
  language?: CreateBatchBodyLanguage;
  backgroundColor?: string;
  logoUrl?: string;
  footerText?: string;
  expiresAt?: Date;
}